requirements.txt
Django>=5.0,<5.1
python-decouple
groq